Move BraceMatching data service down to Features layer.
