+++ src/EditorFeatures/Core/ExternalAccess/VSTypeScript/VSTypeScriptBraceMatcher.cs
@@ -3,16 +3,17 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.ComponentModel.Composition;
+using System.Composition;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.BraceMatching;
 using Microsoft.CodeAnalysis.Editor;
 using Microsoft.CodeAnalysis.ExternalAccess.VSTypeScript.Api;
 using Microsoft.CodeAnalysis.Host.Mef;
 
 namespace Microsoft.CodeAnalysis.ExternalAccess.VSTypeScript
 {
-    [ExportBraceMatcher(InternalLanguageNames.TypeScript)]
+    [ExportBraceMatcher(InternalLanguageNames.TypeScript), Shared]
     internal sealed class VSTypeScriptBraceMatcher : IBraceMatcher
     {
         private readonly IVSTypeScriptBraceMatcherImplementation _impl;
