+++ src/Features/Core/Portable/BraceMatching/AbstractBraceMatcher.cs
@@ -8,7 +8,6 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Editor;
 
 namespace Microsoft.CodeAnalysis.BraceMatching
 {
