+++ src/Features/Core/Portable/BraceMatching/IBraceMatcher.cs
@@ -5,7 +5,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.CodeAnalysis.Editor
+namespace Microsoft.CodeAnalysis.BraceMatching
 {
     internal interface IBraceMatcher
     {
