+++ src/Features/Core/Portable/BraceMatching/IBraceMatchingServiceExtensions.cs
@@ -6,7 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Microsoft.CodeAnalysis.Editor.Shared.Extensions
+namespace Microsoft.CodeAnalysis.BraceMatching
 {
     internal static class IBraceMatchingServiceExtensions
     {
