+++ src/Features/VisualBasic/Portable/BraceMatching/LessThanGreaterThanBraceMatcher.vb
@@ -2,11 +2,12 @@
 ' The .NET Foundation licenses this file to you under the MIT license.
 ' See the LICENSE file in the project root for more information.
 
-Imports System.ComponentModel.Composition
+Imports System.Composition
+Imports Microsoft.CodeAnalysis.BraceMatching
 Imports Microsoft.CodeAnalysis.Host.Mef
 
-Namespace Microsoft.CodeAnalysis.Editor.VisualBasic.BraceMatching
-    <ExportBraceMatcher(LanguageNames.VisualBasic)>
+Namespace Microsoft.CodeAnalysis.VisualBasic.BraceMatching
+    <ExportBraceMatcher(LanguageNames.VisualBasic), [Shared]>
     Friend Class LessThanGreaterThanBraceMatcher
         Inherits AbstractVisualBasicBraceMatcher
